Teams/channel member

Class representing a member of a team or channel (which will normally be a user in Azure Active Directory).

Format

An R6 object of class ms_team_member, inheriting from ms_object.

Fields

token: The token used to authenticate with the Graph host.

tenant: The Azure Active Directory tenant for the parent object.

type: One of "team member", "channel member" or "chat member" depending on the parent object.

properties: The item properties (metadata).

Methods

new(...): Initialize a new object. Do not call this directly; see 'Initialization' below.

delete(confirm=TRUE): Delete this member.

update(...): Update the member's properties (metadata) in Microsoft Graph.

do_operation(...): Carry out an arbitrary operation on the member.

sync_fields(): Synchronise the R object with the member metadata in Microsoft Graph.

get_aaduser(): Get the AAD information for the member; returns an object of class AzureGraph::az_user.

Initialization

Creating new objects of this class should be done via the get_member and list_members methods of the ms_teamand ms_channel classes. Calling the new() method for this class only constructs the R object; it does not call the Microsoft Graph API to retrieve or create the actual member.

See also